Refactor BurerMonteiro into subfolder
