+++ src/LowRankOpt.jl
@@ -22,7 +22,7 @@ include("model.jl")
 include("buffer.jl")
 include("schur.jl")
 include("MOI_wrapper.jl")
-include("BurerMonteiro.jl")
 include("errors.jl")
+include("BurerMonteiro/BurerMonteiro.jl")
 
 end # module LowRankOpt
